+++ src/help.ts
@@ -1,1 +1,0 @@
-export const supportedUrlLink = "https://github.com/libsql/libsql-client-ts#supported-urls";